aws-java/RDS/rds-demos/rds-with-aws-secretmanager-jdbc/rds-aws-secretmanager-jdbc-demo/src/main/java/com/myapps/rds/service/EmployeeService.java
package com.myapps.rds.service;


import com.myapps.rds.model.Employee;
import com.myapps.rds.repository.EmployeeRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;


@Service
@RequiredArgsConstructor
public class EmployeeService {

    private final EmployeeRepository employeeRepository;


    public Employee createEmployee(Employee employee) {
        return employeeRepository.save(employee);
    }

    public List<Employee> getAll() {
        return employeeRepository.findAll();
    }

}
